+++ crates/zng/src/third_party.rs
@@ -223,7 +223,7 @@ pub(crate) fn setup_default_view() {
                         a.set_view(Stack! {
                             toggle::selector = toggle::Selector::single(selected.clone());
                             direction = StackDirection::top_to_bottom();
-                            children = licenses.into_iter().map(default_item_view).collect::<UiVec>();
+                            children = licenses.into_iter().map(default_item_view);
                         })
                     });
                 };