+++ tests/render-tests/src/tests.rs
@@ -18,15 +18,12 @@ pub async fn bw_rgb(render_mode: RenderMode, scale_factor: Factor) {
         clmv!(colors, || {
             Stack! {
                 direction = StackDirection::left_to_right();
-                children = colors
-                    .iter()
-                    .map(|c| {
-                        Wgt! {
-                            widget::background_color = *c;
-                            layout::size = (5, 10);
-                        }
-                    })
-                    .collect::<UiVec>()
+                children = colors.iter().map(|c| {
+                    Wgt! {
+                        widget::background_color = *c;
+                        layout::size = (5, 10);
+                    }
+                })
             }
         }),
     );